+++ src/functions/appendToLocalStorageArray.js
