+++ src/functions/relativeTimeCalculator.js
