+++ src/pages/FeedPage/components/AnswerCluster.jsx
@@ -4,8 +4,8 @@ import useIntersectionObserver from '@hooks/useIntersectionObserver';
 import styled from 'styled-components';
 
 import AnswerItem from './AnswerItem';
-import Empty from '../../assets/Empty.svg?react';
-import Message from '../../assets/Messages.svg?react';
+import Empty from '../../../assets/Empty.svg?react';
+import Message from '../../../assets/Messages.svg?react';
 
 function AnswerCluster({ subjectInfo, result, questions, callback }) {
   const loadingRef = useRef(null);
