+++ src/pages/FeedPage/components/QuestionBox.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { relativeTimeCalculator } from './relativeTimeCalculator';
+import { relativeTimeCalculator } from '../../../functions/relativeTimeCalculator';
 
 function QuestionBox({ children }) {
   const time = relativeTimeCalculator(children.createdAt);
